25: add a style that more closely matches the Chinese
The Chinese hardsubs seem to be {\fnNoto Sans CJK SC\fs92\fsp2.4},
or more fully, with approximate border and position:

Style: Default,Noto Sans CJK SC,92,&H00FFFFFF,&H00FFFFFF,&H28000000,&H00000000,0,0,0,0,100,100,2.4,0,1,1.75,0,2,44,0,12,0

Thus, for a "perfect" match, English could use any of the following:
  * Noto Sans CJK = Source Han Sans,
  * Source Sans (has the same Latin shapes but different metrics),
  * Noto Sans (has different Latin shapes but is recommended
    by Google in combination with Noto Sans CJK).

Here's an example style based on Noto Sans.
The font size is smaller than for Chinese to make the glyphs' actual size
similar to that in the current Arial-based style, which looks pretty good.
Sadly, the line height (line gap) is much bigger than Arial's,
and it looks bad, so this style isn't set as default in this commit.
Perhaps I'll either modify the font to reduce the line height
or stick to one-liners throughout the script.